autopilot-serviceを4つのモジュールに分割してコードの保守性を向上
## 変更内容
- 973行の大きなファイルを4つの専門的なモジュールに分割
  - autopilot-lifecycle.js (217行): ライフサイクル管理とタイマー
  - autopilot-executor.js (372行): 実行ロジック
  - autopilot-planner.js (296行): トレードルートのプランニング
  - autopilot-report.js (141行): レポート生成とシミュレーション
- autopilot-service.jsを統合モジュールに変更（後方互換性維持）

## 利点
- 各モジュールが単一責任の原則に従う
- テストしやすい小さなモジュール
- 循環依存を避ける依存性注入パターンの使用
- 既存のインポートを変更せずに後方互換性を維持

## テスト
- 全30個のテストが合格
- 既存の機能に影響なし
